requirements.txt
# django~=2.2.10
# djangorestframework==3.11.0
# djangorestframework-jsonapi==3.1.0

django
djangorestframework
djangorestframework-jsonapi

# git+https://github.com/django-json-api/django-rest-framework-json-api
